+++ docs/CHANGELOG.md
@@ -5,10 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
+
+### Added
+
+- Add status page text constants
+
+### Changed
+
+- Switch status page text to use constants
 
 ### Fixed
 
+- Login button flashing when navigating to new tab
 - Confirm email component flashing when navigating to home page
 
 ## [2.0.3](https://github.com/hackmcgill/dashboard/tree/2.0.3) - 2019-12-28